Further review comments
- Move all CSV endpoints to a new csv.py file, consolidate with csvcommon.py
- Remove the get() method added to the model
