+++ app/api/__init__.py
@@ -9,3 +9,4 @@ api = Blueprint('api', __name__)
 
 from app.api import data
 from app.api import public
+from app.api import csv
